requirements.txt
# Core Dependencies
google-cloud-bigquery==3.14.1
google-cloud-bigquery-storage==2.24.0
db-dtypes==1.2.0

# Authentication
google-auth==2.25.2
google-auth-oauthlib==1.2.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
tqdm==4.66.1
colorama==0.4.6

# Optional: For advanced features
# bigframes==0.20.0  # Uncomment for BigFrames support
